+++ verl/workers/engine/megatron/transformer_impl.py
@@ -640,7 +640,7 @@ class MegatronEngineWithLMHead(MegatronEngine):
             # avoid non-positive temperature such as padding
             temperature[temperature <= 0] = 1e-8
             assert torch.all(temperature > 0).item(), f"temperature tensor must be positive. Got {temperature}"
-            logits.div_(temperature.unsqueeze(dim=-1))
+            logits.div_(temperature.unsqueeze(dim=-1).to(logits.dtype))
             ret = {}
             if calculate_entropy:
                 logits_bak = logits.clone()
